+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `compare.py` create new `assert_within_tolerance` function to use file-specific tolerances
 * `autorift_golden.json.j2` includes L5+5, L7+7, L7+8, and L8+7 pairs in different projections to test reprojection code
 * `--user-id` pytest CLI argument to allow finding products submitted by a different user than the authorized user
+* `test_opera_rtc_s1` test suite
 
 ### Added
 * `test_autorift.py` golden test for the autoRIFT plugin